Test every node in the cluster in BroadcastTest
We were using threads to access the socket pool, which only has sockets
for a single host, so the host was getting reset to the second one, and
we were only broadcasting the command to the final host in the list.
This happened to be caught in the test because we only look at the first
host, but in fact the second one got the command just fine. In order to
avoid future possible broadcast errors where one host gets the command
and the other doesn't, update the test to check both