Separate MySQL Plugin from DB Plugin
These two plugins used to require eachother but somewhere along the way
when we changed plugin defintions we made the MySQL plugin extend the DB plugin so that
the MySQL plugin could respond to its own Query commands. However this breaks the intended internal
usage of the MySQL Plugin because we use a custom separate plugin to respond to Query. This change
brings back this intended functionality and separates the two plugins. It also:

1. DRYs the version response code
2. Fixes returning actual error messages to the client.
3. Adds support for queries Alteryx uses for schema discovery.